[complete and verify 30-min slot generation for doctors
- Finalized GET endpoint for doctor availability. - Confirmed JSON response with active time slots in Postman.]
